ci: backport-verification test (#3385)
Due to nature of NGO of having 2 development branches (`develop` for
NGOv1.X and `develop-2.0.0` for NGOv2.X) we need to ensure that relevant
changes are landing in only one branch (when it's relevant to only
develop-2.0.0 for example) or both of them.

For people not familiar with NGO structure (but also for us) it might be
easy to omit the fact of making a backport (or forward port) so this PR
includes a GitHub workflow that will check if "**## Backport**" section
exist in the PR thus making it mandatory. This will add a bit of work to
each PR but it will ensure that developers get reminded about backports
and hopefully this will lead to

- Easier PR tracking since every PR will contain a description of a
backport or reasoning why a backport is not needed
- Reduction of potential forgotten backports which in turn could
increase overall quality of NGO

I tested this workflow on my fork of NGO repo and seems to be working as
expected

## Backport
Backport created in
https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3393.
After this PR gets approved I will modify branch protection rules to
block PRs from merging if this check fails.

---------

Co-authored-by: Emma <emma.mcmillan@unity3d.com>