docs: [2.X] fixes of PVP exceptions (round 2) (#3375)
This PR is a continuation of
https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3225
and focuses on fixing PVP errors related to PVP-151-1 "**_Public APIs
should be documented_**" and all remaining PVP errors **present in
pvpExceptions.json file**

We currently have 137 errors to address so this PR will be a
collaborative effort. If anyone want's to add to it what we should do
is:

1. Check errors present in pvpExceptions.json file (preferably under
"PVP-150-1", choose some to fix
2. Address those errors and when making commit remove the fixed errors
from pvpExceptions file

In general [new CI
implementation](https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3193)
will catch all new errors like this so all we need to do is fix the
present ones. This PR **does not** focus on PVP present in gold profile
but if capacity allows, those can also be addressed.

**Error present in "PVP-41-1" is to be expected** so there is no need of
fixing this one!
Another thing is that we could ignore error related to missing APIs for
tests (so TestHelpers, EditorTests, RuntimeTests) because those are not
affecting users in any way.

## Backport 
This PR has its equivalent in the sense of
https://github.com/Unity-Technologies/com.unity.netcode.gameobjects/pull/3374