fix: distributable or sessionowner flag always defaults for in-scene placed NetworkObjects (#3407)
This PR removes the opinionated restrictions to permissions with
in-scene placed NetworkObjects and matches the DAHost redistribution
logic to mirror the rust server's when a client disconnects.

[MTT-11973](https://jira.unity3d.com/browse/MTT-11973)

## Changelog

- Fixed: Issue where root level in-scene placed `NetworkObject`s would
only allow the ownership permission to be no less than distributable or
sessionowner.

## Testing and Documentation

- No tests have been added (requires manually testing in-editor)
- No documentation changes or additions were necessary.

<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->

## Backport

No backport is needed since this is a distributed authority specific
fix.