fix: new interpolators delaying between stopping and starting (#3413)
This fixes the issue where a server authority instance would lag behind
when using the new interpolators. This was especially visible when the
moving body completely stopped, all state updates were processed, and
then shortly later it began to move again there would be a delay between
the 1st state update and the rest of the pending state updates.

This also fixes an issue where if the frame rate takes longer than the
tick frequency authority instances could invoke the same check for state
changes more than once in the same frame. This is not needed since
changes to the transform are frame driven and the check should only be
performed once per new tick(s).

<!-- Add short version of the JIRA ticket to the PR title (e.g. "feat:
new shiny feature [MTT-123]") -->

## Changelog

- Fixed: Issue where the authority instance of `NetworkTransform` could
check for state updates more than one time in a frame if the frame rate
is slower than the tick frequency.
- Fixed: Issue where the new interpolator types were blocking after the
first consumption of a sequence of buffered state updates.

## Testing and Documentation

- Includes the integration test: `InterpolationStopAndStartMotionTest`.
- No documentation changes or additions were necessary.

<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->

## Backport

No backport is needed for this update since it pertains to 2.x only
features.