fix
Fixing an issue where trying to parse for a bool I did not verify the alternative of no environment variable being set and the string value being "unset" as opposed to "false".
This fixes that issue and if it is not a valid bool string the it logs a warning and disables CMB service testing.