fix: NetworkVariables with NetworkVariableUpdateTraits can cause other NetworkVariables to drop changes (up-port) (#3466)
This PR resolves an issue where a `NetworkBehaviour` with multiple
`NetworkVariables` could drop changes if one of the `NetworkVariables`
has unique `NetworkVariableUpdateTraits` set, is dirty, but is not ready
to send.

Based on user @khyperia's submission #3462.

## Changelog

- Fixed: issue where `NetworkVariable`s on a `NetworkBehaviour` could
fail to synchronize changes if one has `NetworkVariableUpdateTraits` set
and is dirty but is not ready to send.

## Testing and Documentation

- Includes integration test
`NetworkVariableTraitsTests.WhenNonTraitsIsDirtyButTraitsIsNotReadyToSend`.
- No documentation changes or additions were necessary.

## Backport

This is an up-port of #3465

---------

Co-authored-by: khyperia <953151+khyperia@users.noreply.github.com>