fix: disabling Physics2D or Physics compilation errors (#3422)
This PR resolves the issue where disabling the physics or physics2d
package module while leaving the other still enabled could cause
compilation issues.

## Changelog

- Fixed: Issue where disabling the physics or physics2D package modules
could result in a compilation error.

## Testing and Documentation

- No tests have been added. (has to be manually tested)
- No documentation changes or additions were necessary.

<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->

## Backport
No backport is needed since this is a v2.x specific issue.
<!-- If this is a backport:
 - Add the following to the PR title: "\[Backport\] ..." .
 - Link to the original PR.
If this needs a backport - state this here
If a backport is not needed please provide the reason why.
If the "Backports" section is not present it will lead to a CI test
failure.
-->