feat: add singleplayer transport component (#3473)
This PR includes the `SinglePlayerTransport` that provides users with
the ability to start a single player game if:
- using either network topology
- using the `SinglePlayerTransport` 
- starting a session as a host 

<!-- Add short version of the JIRA ticket to the PR title (e.g. "feat:
new shiny feature [MTT-123]") -->

## Changelog

- Added: `SinglePlayerTransport` that provides the ability to start as a
host for a single player network session.

## Testing and Documentation

- Includes SinglePlayerTransportTests integration tests.
- Requires adding section to public documentation
([PR-1476](https://github.com/Unity-Technologies/com.unity.multiplayer.docs/pull/1476)).


## Backport

Backported via #3475

<!-- If this is a backport:
 - Add the following to the PR title: "\[Backport\] ..." .
 - Link to the original PR.
If this needs a backport - state this here
If a backport is not needed please provide the reason why.
If the "Backports" section is not present it will lead to a CI test
failure.
-->