fix: don't destroy with owner not being honored when client exits (#3477)
Based on issue #3472, this PR resolves the issue where the
`NetworkObject.DontDestroyWithOwner` was not being honored.

<!-- Add short version of the JIRA ticket to the PR title (e.g. "feat:
new shiny feature [MTT-123]") -->

## Changelog

- Fixed: Issue where the `NetworkObject.DontDestroyWithOwner` was not
being honored.

## Testing and Documentation

- Includes integration test
`NetworkObjectDontDestroyWithOwnerTests.NetworkShowThenClientDisconnects`.
- No documentation changes or additions were necessary.


<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->

## Backport
The v1.x branch does no have this issue, but went ahead and backported
the test in #3478.

<!-- If this is a backport:
 - Add the following to the PR title: "\[Backport\] ..." .
 - Link to the original PR.
If this needs a backport - state this here
If a backport is not needed please provide the reason why.
If the "Backports" section is not present it will lead to a CI test
failure.
-->