fix
This fixes the issue where spawning a prefab that is configured to be spawned with no observers just prior to synchronizing a client and showing the spawned object to the client about to be synchronized within a NetworkBehaviour attached to the newly spawn prefab instance would result in a duplicate creation of the same prefab instance due to NetworkShow being deferred until the end of the frame.

This fix excludes any spawned NetworkObjects that have pending visibility for a client that is being synchronized.