fix: potential small memory leak in ConncectionApprovedMessage (#3486)
This resolves an issue where a very small memory leak could occur within
the `ConnectionApprovedMessage`.

## Changelog

- Fixed: Issue where there was a potential for a small memory leak in
the `ConnectionApprovedMessage`.

## Testing and Documentation

- No tests have been added.
- No documentation changes or additions were necessary.

<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->

## Backport

No backport is needed as this fix is specific to v2.x.

<!-- If this is a backport:
 - Add the following to the PR title: "\[Backport\] ..." .
 - Link to the original PR.
If this needs a backport - state this here
If a backport is not needed please provide the reason why.
If the "Backports" section is not present it will lead to a CI test
failure.
-->