chore: merge 2.4.1 back into develop-2.0.0 (#3496)
Merging the v2.4.1 changes back into develop-2.0.0.
_(merge when published)_


## Changelog

NA

## Testing and Documentation

- No tests have been added.
- No documentation changes or additions were necessary.

<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->

## Backport

This PR is specific to v2.x branch. No backport needed.

<!-- If this is a backport:
 - Add the following to the PR title: "\[Backport\] ..." .
 - Link to the original PR.
If this needs a backport - state this here
If a backport is not needed please provide the reason why.
If the "Backports" section is not present it will lead to a CI test
failure.
-->

---------

Co-authored-by: Michał Chrobot <124174716+michalChrobot@users.noreply.github.com>
Co-authored-by: michalChrobot <michal.chrobot@unity3d.com>