fix: DestroyWithScene causes errors on non-authority side (#3500)
### Description

This PR fixes the issue where enabling the
`NetworkObject.DestroyWithScene` would cause errors on non-authority
instances when loading a new scene (in single mode) or unloading the
scene that the `NetworkObject` instance currently resides and using the
distributed authority network topology.

Fix: #3144

[MTTB-1384](https://jira.unity3d.com/browse/MTTB-1384)

## Changelog

- Fixed: Distributed authority related issue where enabling the
`NetworkObject.DestroyWithScene` would cause errors when a destroying
non-authority instances due to loading (single mode) or unloading scene
events.

## Testing and Documentation

- Includes new integration test: `NetworkObjectDestroyWithSceneTests`.
- No documentation changes or additions were necessary.

<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->

## Backport

This is a v2.x distributed authority only issue.

<!-- If this is a backport:
 - Add the following to the PR title: "\[Backport\] ..." .
 - Link to the original PR.
If this needs a backport - state this here
If a backport is not needed please provide the reason why.
If the "Backports" section is not present it will lead to a CI test
failure.
-->