fix: DontDestroyWithOwner v2.4.0 - v2.4.2 client-server regression (#3522)
This resolves the
[Forum-1658887](https://discussions.unity.com/t/network-objects-dont-despawn-on-disconnect-even-with-dont-destroy-with-owner-unchecked/1658887)
discovered client-server specific issue in the v2.4.2 release where
spawned objects with the default `NetworkObject.DontDestroyWithOwner`
setting (disabled/false) would not be despawned and destroyed when the
owning client disconnected.


## Changelog

- Fixed: Issue where spawned objects with
`NetworkObject.DontDestroyWithOwner` set to `false` would not be
destroyed when using a client-server network topology.

## Testing and Documentation

- Includes updates to the `NetworkObjectDontDestroyWithOwnerTests`
integration test.
- No documentation changes or additions were necessary.

<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->

## Backport
This is specific to NGO v2.4.2 release.

<!-- If this is a backport:
 - Add the following to the PR title: "\[Backport\] ..." .
 - Link to the original PR.
If this needs a backport - state this here
If a backport is not needed please provide the reason why.
If the "Backports" section is not present it will lead to a CI test
failure.
-->