feat: Make INetworkStreamDriverConstructor useful (#3501)
This PR addresses MTTB-1331 by making it simpler to use a custom driver
constructor without breaking things.

`INetworkStreamDriverConstructor` is a mechanism that was added a long
time ago that allows users to override the creation of the
`NetworkDriver` that `UnityTransport` is using. As the name might hint
at, it was pulled directly out of N4E and was never properly integrated
into `UnityTransport`.

For example, before this PR, a user implementing a custom driver
constructor would need to recreate the settings and pipelines manually
if they wanted to mimic the default behavior of `UnityTransport`, which
required reading the code to figure out how we are settings things up.
In the case of pipelines, it's not even possible to do that in all cases
because one of the pipelines used when the multiplayer tools package is
installed is internal.

This PR addresses this by making both the default settings and default
pipeline configurations available through the new
`GetDefaultNetworkSettings` and `GetDefaultPipelineConfigurations`
methods. A user can use these methods to reuse or extend the
configuration `UnityTransport` uses by default. The (new and improved)
documentation of the driver constructor interface provides an example of
how to do so.

As a bonus, this centralizes all the `NetworkSettings` in one place
which makes the code a tiny bit more organized.

## Changelog

- Added: Added methods `GetDefaultNetworkSettings` and
`GetDefaultPipelineConfigurations` to `UnityTransport`. These can be
used to retrieve the default settings and pipeline stages that are used
by `UnityTransport`. This is useful when providing a custom driver
constructor through `UnityTransport.s_DriverConstructor`, since it
allows reusing or tuning the existing configuration instead of trying to
recreate it. This means a transport with a custom driver can now easily
benefit from most of the features of `UnityTransport`, like integration
with the Network Simulator and Network Profiler from the multiplayer
tools package.

## Testing and Documentation

- No tests have been added.
- Includes edits to existing public API documentation.

## Backport

I don't plan on backporting this, unless the team feels strongly
otherwise.

---------

Co-authored-by: Noel Stephens <noel.stephens@unity3d.com>