fix: NetworkObject.GetNetworkBehaviourAtOrderIndex permission level (#3541)
This resolves a regression bug where
NetworkObject.GetNetworkBehaviourAtOrderIndex was converted from public
to internal during the original v2.0.0-pre branch creation.

MTTB-1427
fix: #3528

<!-- Add short version of the JIRA ticket to the PR title (e.g. "feat:
new shiny feature [MTT-123]") -->

## Changelog

- Fixed: Regression issue in v2.x where
`NetworkObject.GetNetworkBehaviourAtOrderIndex` was converted from
public to internal.

## Testing and Documentation

- No tests have been added.
- No documentation changes or additions were necessary.

<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->

## Backport
This issue is specific to v2.x.

<!-- If this is a backport:
 - Add the following to the PR title: "\[Backport\] ..." .
 - Link to the original PR.
If this needs a backport - state this here
If a backport is not needed please provide the reason why.
If the "Backports" section is not present it will lead to a CI test
failure.
-->