docs: Applying docs fixes for patch release (#3535)
This PR includes the following:

- Various link fixes identified on another pass
- Reinstating the sample docs until we can find a better home for them
(and moving them under their own samples folder)

It _does not_ include a fix for the broken external code links. These
aren't supported by the package docs tooling, and we need to discuss how
we're going to get around the issue before we merge/release the docs.

## Changelog

- Fixed: Various small issues with documentation.

## Testing and Documentation

- No tests have been added.
- Includes edits to existing public manual documentation.

## Backport

#3536