fix: Extended UniversalRpcTests run times (#3560)
This PR condenses the total number of tests from roughly ~10k down to
~3.5k by grouping like test types and migrating all `[Values]` decorated
parameters to local properties.

**This helps to reduce:**
- The time spent setting up and tearing down a test for each unique
parameter value combination.
- The time spent starting and stopping NetworkManagers.
- The total number of tests is more than cut in half.
- _There is a known limitation with 2021 & 2022 TestRunner where
reaching/exceed the 8-10k total tests count range can start to impact
performance._

## Changelog

NA

## Testing and Documentation

- Includes some `UniversalRpcTests` refactoring.
- Includes new `UniversalRpcGroupedTests`.
- Includes commenting out all legacy tests that were migrated into
`UniversalRpcGroupedTests`.

<!-- Uncomment and mark items off with a * if this PR deprecates any
API:
### Deprecated API
- [ ] An `[Obsolete]` attribute was added along with a `(RemovedAfter
yyyy-mm-dd)` entry.
- [ ] An [api updater] was added.
- [ ] Deprecation of the API is explained in the CHANGELOG.
- [ ] The users can understand why this API was removed and what they
should use instead.
-->

## Backport
This most likely could be up-ported to v2.x, but for the time being this
should be reviewed merged for the v1.14.0 release.

<!-- If this is a backport:
 - Add the following to the PR title: "\[Backport\] ..." .
 - Link to the original PR.
If this needs a backport - state this here
If a backport is not needed please provide the reason why.
If the "Backports" section is not present it will lead to a CI test
failure.
-->